Created base level + Dynamic Room Generation
Rooms are dynamically generated. they consist of random floor tiles, and
around each room there are random wall tiles. Inside the rooms, there
are enemies, items and inner walls scattered around randomly.

the range at which the width and height of the rooms generate can also
be altered.

Note, the connection to the rooms are not implemented yet.

When you run the scene, be sure to check it out through the scene view,
as opposed to the in game camera view.

Also, variables can be changed in the game manager script, located in
the gamemanager gameobject.