Added Connections to Rooms + Player Starting Point + Camera Change
Added:

- Connections to Rooms
- Player starts in a random room within the map
- Camera zoomed in a little bit, and less rooms are generated (to make
it easier to see what's going on.) This can be tweaked easily.

Note: the rooms connect, but they don't overwrite the walls tat are
already there. Because of that, I removed the collision on the walls, so
that the player can pass. I also commented out the 1 line in the code
that spawns the enemies, since the enemies would be running around
everywhere otherwise.